Mark files_trashbin/trashbin::testExpireOldFilesShared as incomplete on Windows
We can not test user1's filesystem anymore, because the StorageMapper
purges the Mapper table upon destroy(). Since we logged in a second user
the Mapper does not have the root entry `test-trashbin-user1/files_trashbin`
anymore. Therefor instead of `files-trashbin` it now tries to use `files_trashbin`
and therefor fails to strip the path correctly, so the file names are empty,
which later throws an error when trying to get the file extension.

This should be fixed, once owncloud/core#11233 is merged, since it makes the
Mapper recursive and generates any missing parents.